requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
twilio
openai
faster-whisper
soundfile
numpy
httpx